Create task 25.py file
Task № 25:
Write a program that reads the coordinates of a point and
the radius of a circle and determines whether the point falls inside a circle
with a given radius, the center of which is at the center of coordinates.

Print "YES" or "NO" as a result (in capital letters, without quotes).